# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Metacognitive Interview API
  version: 0.1.0
  description: |
    MVP + 扩展版接口契约（FastAPI/OpenAPI 3.0）。
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
...
  /sessions/{session_id}/annotations:
    post:
      tags: [annotation]
      summary: Submit human annotation for one turn (calibration)
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanAnnotationCreateRequest"
      responses:
        "200":
          description: Annotation stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAnnotationCreate"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/question_sets/{question_set_id}:
    get:
      tags: [admin]
      summary: Get question set detail (tree + probes + perturbations)
      parameters:
        - name: question_set_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question set detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseQuestionSetGet"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/rubrics/{rubric_id}:
    get:
      tags: [admin]
      summary: Get rubric detail
      parameters:
        - name: rubric_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rubric detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRubricGet"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_ARGUMENT
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
          nullable: true
    ApiResponseError:
      type: object
      required: [ok, data, error, trace_id]
      properties:
        ok:
          type: boolean
          example: false
        data:
          nullable: true
        error:
          $ref: "#/components/schemas/ApiError"
        trace_id:
          type: string
...
    DimScores:
      type: object
      required: [plan, monitor, evaluate, adapt]
      properties:
        plan:
          type: number
        monitor:
          type: number
        evaluate:
          type: number
        adapt:
          type: number
    EvidenceSpan:
      type: object
      required: [dimension, quote, reason]
      properties:
        dimension:
          type: string
          enum: [plan, monitor, evaluate, adapt]
        quote:
          type: string
        start:
          type: integer
          nullable: true
        end:
          type: integer
          nullable: true
        reason:
          type: string
...
    HumanAnnotationCreateRequest:
      type: object
      required: [turn_id, human_scores]
      properties:
        turn_id:
          type: string
        human_scores:
          $ref: "#/components/schemas/DimScores"
        notes:
          type: string
          nullable: true
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceSpan"
          nullable: true

    # ========== Admin objects (read-only MVP) ==========
...
    QuestionSetDetail:
      type: object
      required: [question_set_id, title, questions]
      properties:
        question_set_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        questions:
          type: array
          items:
            type: object
            additionalProperties: true
...
    RubricDetail:
      type: object
      required: [rubric_id, title, scale]
      properties:
        rubric_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        scale:
          type: object
          additionalProperties: true

    # ========== API Responses (typed) ==========
...
    ApiResponseAnnotationCreate:
      type: object
      required: [ok, data, error, trace_id]
      properties:
        ok:
          type: boolean
        data:
          type: object
          required: [stored]
          properties:
            stored:
              type: boolean
              example: true
        error:
          nullable: true
        trace_id:
          type: string
...
    ApiResponseQuestionSetGet:
      type: object
      required: [ok, data, error, trace_id]
      properties:
        ok:
          type: boolean
        data:
          type: object
          required: [question_set]
          properties:
            question_set:
              $ref: "#/components/schemas/QuestionSetDetail"
        error:
          nullable: true
        trace_id:
          type: string
...
    ApiResponseRubricGet:
      type: object
      required: [ok, data, error, trace_id]
      properties:
        ok:
          type: boolean
        data:
          type: object
          required: [rubric]
          properties:
            rubric:
              $ref: "#/components/schemas/RubricDetail"
        error:
          nullable: true
        trace_id:
          type: string